Extract exceptions from file
